+++ src/AngleSharp.Css/Constants/InitialValues.cs
@@ -90,7 +90,7 @@ namespace AngleSharp.Css
         public static readonly ICssValue AnimationNameDecl = new CssConstantValue<Object>(CssKeywords.None, null);
         public static readonly ICssValue AnimationDurationDecl = CssTimeValue.Zero;
         public static readonly ICssValue AnimationTimingFunctionDecl = CssCubicBezierValue.Ease;
-        public static readonly ICssValue AnimationDelayDecl = CssTimeValue.Zero;
+        public static readonly ICssValue AnimationDelayDecl = new CssTimeValue(0, CssTimeValue.Unit.S);
         public static readonly ICssValue AnimationIterationCountDecl = new CssLengthValue(1, CssLengthValue.Unit.None);
         public static readonly ICssValue AnimationDirectionDecl = new CssConstantValue<AnimationDirection>(CssKeywords.Normal, AnimationDirection.Normal);
         public static readonly ICssValue AnimationFillModeDecl = new CssConstantValue<AnimationFillStyle>(CssKeywords.None, AnimationFillStyle.None);